[MOVE/#20] GetDdayScheduleResponse DTO 패키지 이동
